# pkg/api/docs/swagger.yaml
basePath: /
definitions:
...
  api.TokenValidationResponse:
    properties:
      expires_at:
        type: string
      token_name:
        type: string
    type: object
host: localhost:9898
info:
  contact:
    name: Source Code
    url: https://github.com/jojomomojo/poginfo
  description: Go microservice template for Kubernetes.
  license:
    name: MIT License
    url: https://github.com/jojomomojo/poginfo/blob/master/LICENSE
  title: Podinfo API
  version: "2.0"
paths:
...
  /token/validate:
    post:
      consumes:
      - application/json
      description: validates the JWT token
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.TokenValidationResponse'
        "401":
          description: Unauthorized
          schema:
            type: string
      summary: Validate JWT token
      tags:
      - HTTP API
...
schemes:
- http
- https
swagger: "2.0"